my_menu_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.0.3

# Database & Migrations
alembic==1.12.1
psycopg2-binary==2.9.9

# Cache & Redis
redis==5.0.1
aioredis==2.0.1

# Monitoring & Observability
prometheus-client==0.19.0
loguru==0.7.2
psutil==5.9.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Upload & Image Processing
pillow==10.1.0
aiofiles==23.2.0

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0
factory-boy==3.3.0